requirements.txt
# Core
django>=5.1,<5.2
djangorestframework>=3.15,<4
psycopg2-binary>=2.9,<3
django-environ>=0.11,<1

# Server
gunicorn>=22,<23

# Async tasks
celery>=5.4,<6
redis>=5,<6
django-celery-beat>=2.6,<3

# Storage
django-storages>=1.14,<2
boto3>=1.34,<2
Pillow>=10,<11

# Filtering
django-filter>=24,<25

# GIS
GDAL==3.10.3
djangorestframework-gis>=1.0,<2

# API docs
drf-spectacular>=0.27,<1

# Models
django-mptt>=0.16,<1

# Dev / Test
pytest>=8,<9
pytest-django>=4.8,<5
ruff>=0.4,<1
mypy>=1.10,<2
django-stubs>=5,<6
djangorestframework-stubs>=3,<4
pre-commit>=3,<4
